requirements.txt
tkcalendar==1.6.1
python-bidi==0.4.2
arabic-reshaper==3.0.0
